\name{example}
\alias{example}
\title{Example dataset}
\description{
Example dataset for RsqMed.
}
\format{
  Contains the following objects:
  \describe{
    \item{Y}{a vector of 100 observations, representing outcome.}
    \item{X}{a vector of 100 observations, representing independent variable.}
    \item{M}{a 100 * 100 matrix, representing mediators.}
    \item{Cov}{a 100 * 2 matrix, representing covariates.}
  }
}


